split scripts into separate files
